+++ trunk/src/PIC18-new/image.h
@@ -7,8 +7,8 @@
 typedef struct Image
 {
     uint32 address;
-    uint8  width;
-    uint8  height;
+    uint16 width;
+    uint16 height;
     bool   transparent;
 } Image;
 
@@ -18,11 +18,11 @@ typedef struct Point
     int16 y;
 } Point;
 
-Image makeImage(uint32 address, uint8 width, uint8 height);
-Image* createImage(uint32 address, uint8 width, uint8 height);
+Image makeImage(uint32 address, uint16 width, uint16 height);
+Image* createImage(uint32 address, uint16 width, uint16 height);
 
-Image makeOpaqueImage(uint32 address, uint8 width, uint8 height);
-Image* createOpaqueImage(uint32 address, uint8 width, uint8 height);
+Image makeOpaqueImage(uint32 address, uint16 width, uint16 height);
+Image* createOpaqueImage(uint32 address, uint16 width, uint16 height);
 
 void destroyImage(Image* image);
